Move DummyTranslator to decidim-dev
