+++ decidim-core/spec/jobs/decidim/machine_translation_fields_job_spec.rb
@@ -15,7 +15,7 @@ module Decidim
       end
 
       it "calls DummyTranslator to create machine translations" do
-        expect(Decidim::DummyTranslator)
+        expect(Decidim::Dev::DummyTranslator)
           .to receive(:new)
           .with(
             process,
