+++ decidim-dev/spec/lib/dummy_translator_spec.rb
